backend/foodgram/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0 